chore: convert to options struct
